Docker package renaming.
