+++ cmd/logs.go
@@ -20,7 +20,7 @@ func init() {
 		},
 		// Read logs from a container and copy them to stdout
 		Run: func(cmd *cobra.Command, args []string) {
-			c := docker.NewDockerClientOrExit(args[0])
+			c := docker.NewContainerOrExit(args[0])
 
 			tail, err := cmd.Flags().GetInt("tail")
 			if err != nil {
