+++ internal/docker/docker.go
@@ -4,7 +4,11 @@ import (
 	"context"
 	"fmt"
 	"log"
+	"strconv"
 	"strings"
+
+	"github.com/docker/docker/api/types/container"
+	"github.com/docker/go-connections/nat"
 
 	docker "github.com/docker/docker/api/types"
 	"github.com/docker/docker/client"
@@ -18,6 +22,91 @@ func Client() *client.Client {
 	}
 
 	return c
+}
+
+// RunContainer creates a new craft server container and returns a docker client for it.
+// It is the equivalent of the following docker command:
+//
+//    docker run -d -e EULA=TRUE -p <HOST_PORT>:19132/udp <IMAGE_NAME>
+func RunContainer(hostPort int, name string) (*Container, error) {
+	if hostPort == 0 {
+		hostPort = nextAvailablePort()
+	}
+
+	c, err := client.NewClientWithOpts(client.FromEnv, client.WithAPIVersionNegotiation())
+	if err != nil {
+		log.Fatalf("Error: Failed to create new docker client: %s", err)
+	}
+
+	ctx := context.Background()
+
+	// Create port binding between host ip:port and container port
+	hostBinding := nat.PortBinding{
+		HostIP:   anyIP,
+		HostPort: strconv.Itoa(hostPort),
+	}
+
+	containerPort, err := nat.NewPort(protocol, strconv.Itoa(defaultPort))
+	if err != nil {
+		return nil, fmt.Errorf("creating container port: %s", err)
+	}
+
+	portBinding := nat.PortMap{containerPort: []nat.PortBinding{hostBinding}}
+
+	// Request creation of container
+	createResp, err := c.ContainerCreate(
+		ctx,
+		&container.Config{
+			Image: imageName,
+			Env:   []string{"EULA=TRUE"},
+			ExposedPorts: nat.PortSet{
+				containerPort: struct{}{},
+			},
+			AttachStdin:  true,
+			AttachStdout: true,
+			AttachStderr: true,
+			Tty:          true,
+			OpenStdin:    true,
+		},
+		&container.HostConfig{
+			PortBindings: portBinding,
+			AutoRemove:   true,
+		},
+		nil, nil, name,
+	)
+	if err != nil {
+		return nil, fmt.Errorf("creating docker container: %s", err)
+	}
+
+	// Start the container
+	err = c.ContainerStart(ctx, createResp.ID, docker.ContainerStartOptions{})
+	if err != nil {
+		return nil, fmt.Errorf("starting container: %s", err)
+	}
+
+	d := Container{
+		ContainerAPIClient: c,
+		ContainerName:      name,
+		containerID:        createResp.ID,
+	}
+
+	return &d, nil
+}
+
+// ContainerFromName returns the ID of the container with the given name or an error if that container doesn't exist.
+func ContainerID(name string, client client.ContainerAPIClient) (string, error) {
+	containers, err := client.ContainerList(context.Background(), docker.ContainerListOptions{})
+	if err != nil {
+		return "", fmt.Errorf("listing all containers: %s", err)
+	}
+
+	for _, ctr := range containers {
+		if strings.Trim(ctr.Names[0], "/") == name {
+			return ctr.ID, nil
+		}
+	}
+
+	return "", &ContainerNotFoundError{Name: name}
 }
 
 // ContainerNames returns a slice containing the names of all running containers.
@@ -74,17 +163,17 @@ OUTER:
 	panic("100 ports were not available")
 }
 
-// ActiveServerClients returns a DockerClient for each active server.
-func ActiveServerClients() ([]*DockerClient, error) {
+// ActiveServerClients returns a Container for each active server.
+func ActiveServerClients() ([]*Container, error) {
 	names, err := ContainerNames()
 	if err != nil {
 		return nil, fmt.Errorf("getting server names: %s", err)
 	}
 
-	clients := make([]*DockerClient, len(names))
+	clients := make([]*Container, len(names))
 
 	for i, n := range names {
-		c, err := NewDockerClient(n)
+		c, err := NewContainer(n)
 		if err != nil {
 			return nil, fmt.Errorf("creating client for container '%s': %s", n, err)
 		}
