+++ internal/docker/dockerclient.go
@@ -13,9 +13,6 @@ import (
 	"strings"
 	"time"
 
-	"github.com/docker/docker/api/types/container"
-	"github.com/docker/go-connections/nat"
-
 	docker "github.com/docker/docker/api/types"
 	"github.com/docker/docker/client"
 )
@@ -27,15 +24,15 @@ const (
 	imageName   = "danhaledocker/craftmine:v1.7" // The name of the docker image to use
 )
 
-type DockerClient struct {
+type Container struct {
 	client.ContainerAPIClient
 	ContainerName, containerID string
 }
 
-// NewDockerClientOrExit is a convenience function for attempting to find a docker client with the given name. If not
+// NewContainerOrExit is a convenience function for attempting to find a docker client with the given name. If not
 // found, a helpful error message is printed and the program exits without error.
-func NewDockerClientOrExit(containerName string) *DockerClient {
-	d, err := NewDockerClient(containerName)
+func NewContainerOrExit(containerName string) *Container {
+	d, err := NewContainer(containerName)
 
 	if err != nil {
 		// Container was not found
@@ -51,9 +48,9 @@ func NewDockerClientOrExit(containerName string) *DockerClient {
 	return d
 }
 
-// NewDockerClient returns a new default Docker Container API client. If the given container name doesn't exist an error
+// NewContainer returns a new default Docker Container API client. If the given container name doesn't exist an error
 // of type ContainerNotFoundError is returned.
-func NewDockerClient(containerName string) (*DockerClient, error) {
+func NewContainer(containerName string) (*Container, error) {
 	c, err := client.NewClientWithOpts(client.FromEnv, client.WithAPIVersionNegotiation())
 	if err != nil {
 		log.Fatalf("Error: Failed to create new docker client: %s", err)
@@ -64,7 +61,7 @@ func NewDockerClient(containerName string) (*DockerClient, error) {
 		return nil, err
 	}
 
-	d := DockerClient{
+	d := Container{
 		ContainerAPIClient: c,
 		ContainerName:      containerName,
 		containerID:        id,
@@ -73,93 +70,36 @@ func NewDockerClient(containerName string) (*DockerClient, error) {
 	return &d, nil
 }
 
-// NewContainer creates a new craft server container and returns a docker client for it.
-// It is the equivalent of the following docker command:
-//
-//    docker run -d -e EULA=TRUE -p <HOST_PORT>:19132/udp <IMAGE_NAME>
-func NewContainer(hostPort int, name string) (*DockerClient, error) {
-	if hostPort == 0 {
-		hostPort = nextAvailablePort()
-	}
-
-	c, err := client.NewClientWithOpts(client.FromEnv, client.WithAPIVersionNegotiation())
-	if err != nil {
-		log.Fatalf("Error: Failed to create new docker client: %s", err)
-	}
-
-	ctx := context.Background()
-
-	// Create port binding between host ip:port and container port
-	hostBinding := nat.PortBinding{
-		HostIP:   anyIP,
-		HostPort: strconv.Itoa(hostPort),
-	}
-
-	containerPort, err := nat.NewPort(protocol, strconv.Itoa(defaultPort))
-	if err != nil {
-		return nil, fmt.Errorf("creating container port: %s", err)
-	}
-
-	portBinding := nat.PortMap{containerPort: []nat.PortBinding{hostBinding}}
-
-	// Request creation of container
-	createResp, err := c.ContainerCreate(
-		ctx,
-		&container.Config{
-			Image: imageName,
-			Env:   []string{"EULA=TRUE"},
-			ExposedPorts: nat.PortSet{
-				containerPort: struct{}{},
-			},
-			AttachStdin:  true,
-			AttachStdout: true,
-			AttachStderr: true,
-			Tty:          true,
-			OpenStdin:    true,
-		},
-		&container.HostConfig{
-			PortBindings: portBinding,
-			AutoRemove:   true,
-		},
-		nil, nil, name,
+func (d *Container) CopyFrom(containerPath string) (*tar.Reader, error) {
+	data, _, err := d.CopyFromContainer(
+		context.Background(),
+		d.containerID,
+		containerPath,
 	)
 	if err != nil {
-		return nil, fmt.Errorf("creating docker container: %s", err)
+		return nil, fmt.Errorf("copying data from server at '%s': %s", containerPath, err)
 	}
 
-	// Start the container
-	err = c.ContainerStart(ctx, createResp.ID, docker.ContainerStartOptions{})
-	if err != nil {
-		return nil, fmt.Errorf("starting container: %s", err)
-	}
-
-	d := DockerClient{
-		ContainerAPIClient: c,
-		ContainerName:      name,
-		containerID:        createResp.ID,
-	}
-
-	return &d, nil
+	return tar.NewReader(data), nil
 }
 
-// ContainerFromName returns the ID of the container with the given name or an error if that container doesn't exist.
-func ContainerID(name string, client client.ContainerAPIClient) (string, error) {
-	containers, err := client.ContainerList(context.Background(), docker.ContainerListOptions{})
+func (d *Container) CopyTo(destPath string, tar *bytes.Buffer) error {
+	err := d.CopyToContainer(
+		context.Background(),
+		d.containerID,
+		destPath,
+		tar,
+		docker.CopyToContainerOptions{},
+	)
 	if err != nil {
-		return "", fmt.Errorf("listing all containers: %s", err)
+		return fmt.Errorf("copying files to '%s': %s", destPath, err)
 	}
 
-	for _, ctr := range containers {
-		if strings.Trim(ctr.Names[0], "/") == name {
-			return ctr.ID, nil
-		}
-	}
-
-	return "", &ContainerNotFoundError{Name: name}
+	return nil
 }
 
 // Command runs the given arguments separated by spaces as a command in the bedrock_server process cli.
-func (d *DockerClient) Command(args []string) error {
+func (d *Container) Command(args []string) error {
 	// Attach to the container
 	waiter, err := d.ContainerAttach(
 		context.Background(),
@@ -188,7 +128,7 @@ func (d *DockerClient) Command(args []string) error {
 }
 
 // CommandWriter returns a *net.Conn which streams to the mc server process stdin.
-func (d *DockerClient) CommandWriter() (net.Conn, error) {
+func (d *Container) CommandWriter() (net.Conn, error) {
 	// Attach to the container
 	waiter, err := d.ContainerAttach(
 		context.Background(),
@@ -206,7 +146,7 @@ func (d *DockerClient) CommandWriter() (net.Conn, error) {
 }
 
 // Stop stops the docker container.
-func (d *DockerClient) Stop() error {
+func (d *Container) Stop() error {
 	timeout := time.Duration(10)
 
 	return d.ContainerStop(
@@ -218,7 +158,7 @@ func (d *DockerClient) Stop() error {
 
 // LogReader returns a buffer with the stdout and stderr from the running mc server process. New output will continually
 // be sent to the buffer.
-func (d *DockerClient) LogReader(tail int) (*bufio.Reader, error) {
+func (d *Container) LogReader(tail int) (*bufio.Reader, error) {
 	logs, err := d.ContainerLogs(
 		context.Background(),
 		d.containerID,
@@ -238,7 +178,7 @@ func (d *DockerClient) LogReader(tail int) (*bufio.Reader, error) {
 }
 
 // GetPort returns the port players use to connect to this server
-func (d *DockerClient) GetPort() (int, error) {
+func (d *Container) GetPort() (int, error) {
 	c, err := d.ContainerInspect(context.Background(), d.containerID)
 	if err != nil {
 		return 0, err
@@ -266,34 +206,6 @@ func (d *DockerClient) GetPort() (int, error) {
 	}
 
 	return port, nil
-}
-
-func (d *DockerClient) CopyFromTar(containerPath string) (*tar.Reader, error) {
-	data, _, err := d.CopyFromContainer(
-		context.Background(),
-		d.containerID,
-		containerPath,
-	)
-	if err != nil {
-		return nil, fmt.Errorf("copying data from server at '%s': %s", containerPath, err)
-	}
-
-	return tar.NewReader(data), nil
-}
-
-func (d *DockerClient) CopyToTar(destPath string, tar *bytes.Buffer) error {
-	err := d.CopyToContainer(
-		context.Background(),
-		d.containerID,
-		destPath,
-		tar,
-		docker.CopyToContainerOptions{},
-	)
-	if err != nil {
-		return fmt.Errorf("copying files to '%s': %s", destPath, err)
-	}
-
-	return nil
 }
 
 // ContainerNotFoundError tells the caller that no containers were found with the given name.
